+++ doom64/doomdef.h
@@ -24,7 +24,7 @@
 #define	SCREEN_HT	240
 #define	SCREEN_WD	320
 
-extern u32 cfb[2][SCREEN_WD*SCREEN_HT]; // 8036A000
+extern u16 cfb[2][SCREEN_WD*SCREEN_HT]; // 8036A000
 
 /*============================================================================= */
